A few minor refactors
