+++ Befunge-93.Language/Builders.cs
@@ -205,15 +205,4 @@ namespace Befunge_93 {
         //                return NullObject.Instance;
         //        }
         //}
-
-        internal class Befunge93Command : BaseCommand<Func<InterpreterState, SourceCodeTorus, BaseInterpreterStack, Task>> {
-                internal Befunge93Command(Func<InterpreterState, SourceCodeTorus, BaseInterpreterStack, Task> cmd,
-                        string keyWord)
-                        : base(cmd, keyWord) { }
-
-                protected override async Task InterpretAsync(InterpreterState state) => await Command(state,
-                        state.GetSource<SourceCodeTorus>(), state.GetExecutionEnvironment<BaseInterpreterStack>());
-
-                public override object Clone() => new Befunge93Command(Command, KeyWord);
-        }
 }
